added imports recursively define wf_option_filename

+++ alto/commands/cromwell/run.py
@@ -114,14 +114,18 @@ def submit_to_cromwell(
         else:
             raise Exception("Dependency zip file does not exist or is not given in zip format.")
     elif not is_url:
-        doc = WDL.load(workflow_str)
-        deps = []
-        workflow_dir = os.path.dirname(workflow_str)
-        for d in doc.imports:
-            # TODO add imports recursively
-            imported_path = os.path.join(workflow_dir, d.uri)
-            if os.path.exists(imported_path):
-                deps.append(imported_path)
+        deps = set()
+
+        def add_deps(path):
+            workflow_dir = os.path.dirname(path)
+            for d in WDL.load(path).imports:
+                imported_path = os.path.abspath(os.path.join(workflow_dir, d.uri))
+                if os.path.exists(imported_path):
+                    deps.add(imported_path)
+                    add_deps(imported_path)
+
+        # add imports recursively
+        add_deps(workflow_str)
         if len(deps) > 0:
             tmp_zip_file = tempfile.mkstemp(prefix="alto", suffix=".zip")[1]
             with zipfile.ZipFile(tmp_zip_file, "w", zipfile.ZIP_DEFLATED) as out:
@@ -156,6 +160,7 @@ def submit_to_cromwell(
         json.dump(label_dict, fp)
     files["labels"] = open(wf_label_filename, "rb")
 
+    wf_option_filename = None
     # Process job's workflow options.
     if no_cache:
         wf_option_dict = {
@@ -179,7 +184,7 @@ def submit_to_cromwell(
             os.remove(tmp_zip_file)
         if os.path.exists(wf_label_filename):
             os.remove(wf_label_filename)
-        if os.path.exists(wf_option_filename):
+        if wf_option_filename is not None and os.path.exists(wf_option_filename):
             os.remove(wf_option_filename)
 
     # Process response.
